requirements.txt
streamlit==1.40.0
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
plotly==5.24.1
google-cloud-bigquery==3.26.0
google-auth==2.35.0
db-dtypes==1.3.0